fix: #3717 - runtime error in jans-config-api
